No longer use TERM=xterm-256color inside tmux
It breaks most of the ncurses-based apps (like htop or mutt) and should not be used!
As the FAQ says, we should use "screen-256color" from inside tmux, which breaks xterm keys
for vim. So adding a xtermkeys.vim plugin too which tells to vim how to use these xterm keys